* 'master' of github.com:tensorflow/tensorflow: (108 commits)
  [tf.data service] Support __iter__ for tf.data service datasets.
  Avoid partially creating/updating cluster when some workers fail during update.
  Treat "test_xla_gpu" as GPU_TEST in "NamedGPUCombination".
  Enable a disable test as it no longer seems flaky.
  TF_LITE_MAYBE_KERNEL_LOG with a check for non-null context
  Remove async runs of cluster update with pending nodes tests.
  Go: Update generated wrapper functions for TensorFlow ops.
  Internal change
  [tf.data] Moving `cardinality` to core API.
  Add hardswish op to hexagon delegate for UInt8/Int8 inputs
  [WhileV2] Fix potential stack overflow when kernels execute inline.
  Use persistent buffer in xtensa_hifimini/fully_connected.
  Disable flaky test on TAP
  Disable flaky ASAN test
  Disable flaky ASAN test
  Disable max_pool_2d_test under TSAN
  Initial checkin of SavedModelAPI C++ header-only API. Tensor, TensorHandle and ConreteFunction::Run will be added in a subsequent change. See RFC https://github.com/tensorflow/community/pull/207.
  Add the scale propagation pass to the pipeline
  Don't propagate ref type if it wasn't originally done
  Clean up dead core.
  ...

+++ tensorflow/python/keras/layers/preprocessing/normalization.py
@@ -101,7 +101,7 @@ class Normalization(CombinerPreprocessingLayer):
     self.count = self._add_state_variable(
         name=_COUNT_NAME,
         shape=(),
-        dtype=dtypes.int32,
+        dtype=dtypes.int64,
         initializer=init_ops.zeros_initializer)
 
     super(Normalization, self).build(input_shape)
